pyproject.toml
[tool.poetry]
name = "clfgraph"
version = "0.1.0"
description = "Experimental architecture combining the best of shallow Neural Networks with Graph Neural Networks. Rust components."
authors = ["snyata <core@synavate.tech>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "clfgraph", from = "clfGraph/src/" }]
repository = "https://github.com/orion-constellation/clfGraph-mojo"


[tool.poetry.dependencies]
python = "^3.10"
ray = "^2.34.0"
redis = "^5.0.8"
types-redis = "^4.6.0.20240819"
uvicorn = "^0.30.6"
stix2 = "^3.0.1"
scikit-learn = "^1.5.1"
torch = "^2.4.0"
isort = "^5.13.2"
wandb = "^0.17.8"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
pyarrow = "^17.0.0"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
pytest = "^8.3.2"
tqdm = "^4.66.5"
streamlit = "^1.38.0"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
seaborn = "^0.13.2"
uuid = "^1.30"
gensim = "^4.3.3"
nltk = "^3.9.1"



[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pylint = "^3.2.6"
black = "^24.8.0"
flake8 = "^7.1.1"
pytest-cov = "^5.0.0"
vulture = "^2.11"
tree = "^0.2.4"
codecov = "^2.1.13"
poetry-plugin-export = "^1.8.0"
pre-commit = "^3.8.0"
sentry-sdk = "^2.13.0"
huggingface-hub = "^0.24.6"
joblib = "^1.4.2"
datasets = "^2.21.0"
modin = "^0.31.0"
dask = "^2024.8.1"
bokeh = "^3.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
